pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dspy-jira-feedback"
version = "0.1.0"
description = "DSPy-powered Jira issue feedback system with MCP integration"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "dspy-ai>=2.4.0,<3.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "rich>=13.7.0,<15.0.0",
    "httpx>=0.25.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<9.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "pytest-mock>=3.12.0,<4.0.0",
    "black>=23.0.0,<25.0.0",
    "ruff>=0.1.0,<1.0.0",
    "respx>=0.20.0,<1.0.0",
]

[project.scripts]
jira-feedback = "src.app:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
